requirements.txt
pytest==7.4.3
pytest-cov==4.1.0 
pydantic==2.10.6
pydantic[email]
assertpy==1.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.38
alembic==1.14.1
python-dotenv==1.0.0